# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the public pool api from stratum.farm",
        "title": "1oop Pool API",
        "termsOfService": "https://stratum.farm/terms/",
        "contact": {
            "name": "1oop Support",
            "email": "pool@stratum.farm"
        },
        "version": "1.0"
    },
    "host": "152.228.229.130:3000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/pools/{pool_id}/miners/{miner_addr}/workers/{worker_name}": {
            "get": {
                "description": "Get a specific worker from a specific miner from a specific pool",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workers"
                ],
                "summary": "Get a worker",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the pool",
                        "name": "pool_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Address of the miner",
                        "name": "miner_addr",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of the worker",
                        "name": "worker_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.WorkerRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/pools/{pool_id}/miners/{miner_addr}/workers/{worker_name}/performance": {
            "get": {
                "description": "Get the performance from a specific worker from a specific miner from a specific pool",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workers"
                ],
                "summary": "Get performance stats of a worker",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the pool",
                        "name": "pool_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Address of the miner",
                        "name": "miner_addr",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of the worker",
                        "name": "worker_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Specify the sample range (default=day",
                        "name": "perfMode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.WorkerPerformanceRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.PerformanceStats": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "hashrate": {
                    "type": "number"
                },
                "reportedHashrate": {
                    "type": "number"
                },
                "sharesPerSecond": {
                    "type": "number"
                },
                "workersOnline": {
                    "type": "integer"
                }
            }
        },
...
        "api.Worker": {
            "type": "object",
            "properties": {
                "hashrate": {
                    "type": "number"
                },
                "sharesPerSecond": {
                    "type": "number"
                }
            }
        },
        "api.WorkerPerformanceRes": {
            "type": "object",
            "properties": {
                "pageCount": {
                    "type": "integer"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.PerformanceStats"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.WorkerRes": {
            "type": "object",
            "properties": {
                "pageCount": {
                    "type": "integer"
                },
                "result": {
                    "$ref": "#/definitions/api.Worker"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "utils.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        }
    }
}